Added tokenizer module and basic autograd functionality

+++ tensor/include/tensorlib/tokenizer/impl/utf8.h
@@ -2,12 +2,11 @@
 
 #include <cstdint>
 #include <filesystem>
-#include <fstream>
 #include <string_view>
 #include <sys/types.h>
 #include <vector>
 namespace tensorlib::tokenizer::impl::utf8 {
-std::vector<uint32_t> encodeToBytes(std::string_view str);
+std::vector<uint16_t> encodeToBytes(std::string_view str);
 std::string readFile(const std::filesystem::path& data_path);
 
 } // namespace tensorlib::tokenizer::impl::utf8
