+++ tensor/include/tensorlib/tokenizer/trainer.h
@@ -4,8 +4,9 @@
 #include <tensorlib/tokenizer/tokenizer.h>
 
 namespace tensorlib::tokenizer {
+
 class BPETrainer {
   public:
-    static Tokenizer train(const std::filesystem::path& data_path, uint32_t vocab_size);
+    static Tokenizer train(const std::filesystem::path& data_path, uint16_t vocab_size);
 };
 } // namespace tensorlib::tokenizer
