+++ tensor/src/tokenizer/impl/utf8.cpp
@@ -1,11 +1,11 @@
 #include <cstdint>
+#include <fstream>
 #include <string_view>
 #include <tensorlib/tokenizer/impl/utf8.h>
 #include <vector>
-
 namespace tensorlib::tokenizer::impl::utf8 {
-std::vector<uint32_t> encodeToBytes(std::string_view str) {
-    std::vector<uint32_t> bytes(str.begin(), str.end());
+std::vector<uint16_t> encodeToBytes(std::string_view str) {
+    std::vector<uint16_t> bytes(str.begin(), str.end());
     return bytes;
 }
